[BUGFIX] Remove translation of tt_content default values
The default values for tt_content columns defined inside the TsConfig previously localized by the LocalizationUtility.

Since these values are related database values and can contain integer, floats etc. a translation is not the default case.

In order to translate default values for header etc. use the conditions available for the TsConfig.

Refs: #593